Move pure waste-balance core into domain layer
The command core (commands.js) and closing-balance arithmetic
(stream-closing-balance.js) are pure decision logic with no I/O, so they
move from application/ to domain/, beside the WasteBalance model.

Rename foldAggregate to currentWasteBalance: it reads through the stream
repository port, so it is an application read use case and stays in
application/.

The moved domain files still import the shared event-kind, zero-balance and
type vocabulary from repository/stream-schema.js. Lifting that vocabulary
into the domain layer is left to the stream-to-ledger rename, which reworks
that file anyway.
