+++ src/waste-balances/domain/stream-closing-balance.js
